♻️ [Refactoring]: group hooks/ converter implementations under hooks/converter/
Move converter.rs / codex.rs / copilot.rs (and their _test.rs) into
hooks/converter/. Introduce hooks/converter.rs as the sub-parent. Declare
the inner converter module as private (mod converter;, NOT pub mod converter;)
to avoid creating a new public path crate::hooks::converter::converter::*,
and lift the leaf's pub / pub(crate) symbols into the sub-parent's namespace
via pub use self::converter::*; so external paths
crate::hooks::converter::ConversionWarning / SourceFormat / SCRIPTS_DIR
remain unchanged. Apply #[allow(clippy::module_inception)] to the same-name
leaf module (spec-014 commit d6483ff style).

Rewrite cross-subgroup super::* references in the moved files to
super::super::* (e.g. super::super::tool::copilot::CopilotToolMap,
super::super::event::copilot::CopilotEventMap) to make the hierarchy
explicit (ref 論点 b). Same-subgroup references (super::converter::*,
super::codex::*, super::copilot::*) are kept intact. The
use crate::hooks::hook_definition::*; references are kept as-is in this
commit and will be rewritten to super::super::model::hook_definition::*
in the next commit when hook_definition moves into hooks/model/.

Move the #[cfg(test)] mod codex_test; converter_test; copilot_test;
declarations from hooks.rs to the new sub-parent hooks/converter.rs (test
form B). Delete pub(crate) mod codex; and pub(crate) mod copilot; from
hooks.rs; keep pub mod converter; as-is to maintain
crate::hooks::converter::* external exposure.

Refs #247
